combine cli and argparser tests
