+++ .circleci/config.yml
@@ -160,7 +160,7 @@ commonYAMLStructure:
     run:
       name: wait for postgres to be ready
       command: |
-        for i in `seq 1 10`;
+        for i in `seq 1 60`;
         do
           nc -z localhost 5432 && echo Success && exit 0
           echo -n .
@@ -171,7 +171,7 @@ commonYAMLStructure:
     run:
       name: wait for mysql to be ready
       command: |
-        for i in `seq 1 10`;
+        for i in `seq 1 60`;
         do
           nc -z 127.0.0.1 3306 && echo Success && exit 0
           echo -n .
@@ -222,7 +222,6 @@ commonYAMLStructure:
       name: run all tests execluding sql functions tests
       command: |
         . venv/bin/activate
-        python -m tests -v suite argparser
         python -m tests -v suite utils
         python -m tests -v suite cli
   - &run_tests_exc_sql_functions_tests_win
